Add note on and note off

+++ pedalboard/plugins/Chain.h
@@ -104,7 +104,7 @@ inline void init_chain(py::module &m) {
           "process",
           [](std::shared_ptr<Chain> self,
              const py::array_t<float, py::array::c_style> inputArray,
-             const py::array_t<int, py::array::c_style> midiMessages,
+             const py::array_t<float, py::array::c_style> midiMessages,
              double sampleRate, unsigned int bufferSize, bool reset) {
             return process(inputArray, midiMessages, sampleRate, self->getPlugins(),
                            bufferSize, reset);
@@ -119,7 +119,7 @@ inline void init_chain(py::module &m) {
           "process",
           [](std::shared_ptr<Chain> self,
              const py::array_t<double, py::array::c_style> inputArray,
-             const py::array_t<int, py::array::c_style> midiMessages,
+             const py::array_t<float, py::array::c_style> midiMessages,
              double sampleRate, unsigned int bufferSize, bool reset) {
             const py::array_t<float, py::array::c_style> float32InputArray =
                 inputArray.attr("astype")("float32");
